Adding Dockerfile and resources from RHMAP-4570
Taken from danbev's original commit and moved to the public repo

Motivation:
This repository was created to contain the Dockerfile and resources for
Nagios monitoring. This was originally intended to go into the
fh-dockerfiles repository, but it was decided that is should be in it's
own repository instead.

Modifications:
Pulled the changed from RHMAP-4570 and added them to the root of this
project.

JIRA:
https://issues.jboss.org/browse/RHMAP-4821