Add dirs to volume in start script
Motivation:
Things weren't working because we created directories that
were expected to be present in the Dockerfile, but then mounting a
volume over them, so they were hidden.

Modification:
These directories are now created in the start.sh
script, and also in the Dockerfile, so that it shouldn't matter if a
volume is being used or not.

Result:
Things were broken and now they work.