Fix path to sendEmail that nagios uses
Motivation:
Since we started to consume sendEmail from an RPM rather than bundling
it, it's in /usr/bin instead of /usr/local/bin. Nagios was unable to
send mail because of the incorrect location.

Modification:
Fixed the path to the command in the nagios commands that use it.

Result:
Nagios notifications now send email messages as intended